+++ types/loader.d.ts
@@ -1,46 +1,25 @@
 export = loader;
-/** @typedef {import("schema-utils/declarations/validate").Schema} Schema */
-/** @typedef {import("webpack").Compilation} Compilation */
-/**
- * @template T
- * @typedef {Object} LoaderOptions<T>
- * @property {string} [severityError] Allows to choose how errors are displayed.
- * @property {import("./index").Minimizer<T> | import("./index").Minimizer<T>[]} [minimizer]
- * @property {import("./index").Generator<T>[]} [generator]
- */
 /**
  * @template T
  * @this {import("webpack").LoaderContext<LoaderOptions<T>>}
  * @param {Buffer} content
  * @returns {Promise<Buffer | undefined>}
  */
-declare function loader<T>(content: Buffer): Promise<Buffer | undefined>;
-declare class loader<T> {
-  /** @typedef {import("schema-utils/declarations/validate").Schema} Schema */
-  /** @typedef {import("webpack").Compilation} Compilation */
-  /**
-   * @template T
-   * @typedef {Object} LoaderOptions<T>
-   * @property {string} [severityError] Allows to choose how errors are displayed.
-   * @property {import("./index").Minimizer<T> | import("./index").Minimizer<T>[]} [minimizer]
-   * @property {import("./index").Generator<T>[]} [generator]
-   */
-  /**
-   * @template T
-   * @this {import("webpack").LoaderContext<LoaderOptions<T>>}
-   * @param {Buffer} content
-   * @returns {Promise<Buffer | undefined>}
-   */
-  constructor(content: Buffer);
-  resourcePath: string | undefined;
-  resourceQuery: string | undefined;
-}
+declare function loader<T>(
+  this: import("webpack").LoaderContext<LoaderOptions<T>>,
+  content: Buffer
+): Promise<Buffer | undefined>;
 declare namespace loader {
-  export { raw, Schema, Compilation, LoaderOptions };
+  export {
+    raw,
+    Schema,
+    Compilation,
+    WorkerResult,
+    Minimizer,
+    Generator,
+    LoaderOptions,
+  };
 }
-declare var raw: boolean;
-type Schema = import("schema-utils/declarations/validate").Schema;
-type Compilation = import("webpack").Compilation;
 /**
  * <T>
  */
@@ -55,3 +34,15 @@ type LoaderOptions<T> = {
     | undefined;
   generator?: import("./index").Generator<T>[] | undefined;
 };
+declare var raw: boolean;
+type Schema = import("schema-utils/declarations/validate").Schema;
+type Compilation = import("webpack").Compilation;
+type WorkerResult = import("./utils").WorkerResult;
+/**
+ * <T>
+ */
+type Minimizer<T> = import("./index").Minimizer<T>;
+/**
+ * <T>
+ */
+type Generator<T> = import("./index").Generator<T>;